+++ lib/src/easy_dropdown_alignment.dart
@@ -1,0 +1,11 @@
+/// Enumeration representing alignment options for the EasyDropdown widget.
+enum EasyDropdownAlignment {
+  /// Aligns the dropdown to the center relative to the button.
+  center,
+
+  /// Aligns the dropdown to the left edge of the button.
+  left,
+
+  /// Aligns the dropdown to the right edge of the button.
+  right;
+}